code cleanup, updating to Java8 sttyle

+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/ColumnReference.java
@@ -163,8 +163,8 @@ public class ColumnReference extends ValueNode {
 	{
 		this.columnName = (String) columnName;
 		this.tableName = (TableName) tableName;
-		this.setBeginOffset(((Integer) tokBeginOffset).intValue());
-		this.setEndOffset(((Integer) tokEndOffset).intValue());
+		this.setBeginOffset((Integer) tokBeginOffset);
+		this.setEndOffset((Integer) tokEndOffset);
 		tableNumber = -1;
 		remaps = null;
 	}
